Replaces STUB mechanism to support symbolic::Expression in MBP.
